pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "datamd-engine"
version = "0.1.0"
description = "Data Markdown (DataMD) Engine: render .dmd files with embedded data shortcodes"
readme = "README.md"
authors = [{ name = "Kresna Sucandra" }]
license = { text = "MIT" }
requires-python = ">=3.8"
dependencies = [
  "markdown>=3.4.0",
  "pandas>=1.5.0",
  "openpyxl>=3.0.0",
  "xlrd>=2.0.0",
  "odfpy>=1.4.0",
  "pdfplumber>=0.7.0",
  "pytesseract>=0.3.0",
  "Pillow>=9.0.0",
  "moviepy>=1.0.0",
  "tabulate>=0.9.0",
]

[project.scripts]
datamd = "python_implementation.process_dmd:main"

[tool.setuptools]
packages = ["python_implementation"]

[project.optional-dependencies]
watch = [
  "watchdog>=3.0.0",
]
